Updates to install (or upgrade) and run Count.ly Server on macOS

+++ bin/scripts/detect.init.sh
@@ -10,6 +10,9 @@ then
 	if [[ $(/sbin/init --version) =~ upstart ]];
 	then
 	    INITSYS="upstart"
+	elif [ -x /sbin/launchd -a -x /bin/launchctl ];
+	then
+		INITSYS="launchd"
 	fi 2> /dev/null
 else
 	INITSYS="docker" 
@@ -19,6 +22,11 @@ bash "$DIR/commands/$INITSYS/install.sh"
 ln -sf "$DIR/commands/$INITSYS/countly.sh" "$DIR/commands/enabled/countly.sh"
 
 chmod +x "$DIR/commands/countly.sh"
-ln -sf "$DIR/commands/countly.sh" /usr/bin/countly
+if [ -x /usr/bin/csrutil -a "$(/usr/bin/csrutil status | sed -E -e 's/^.+status: *(.+)$/\1/')" == "enabled." ]
+then
+	ln -sf "$DIR/commands/countly.sh" /usr/local/bin/countly
+else
+	ln -sf "$DIR/commands/countly.sh" /usr/bin/countly
+fi
 
 cp -f "$DIR/commands/scripts/autocomplete/countly" /etc/bash_completion.d
